wgpu_quick_start::my_surface

Struct MySurface

pub struct MySurface<'a> { /* private fields */ }

An abstraction of the wgpu::Surface<’lifetime> that gets created.

use wgpu_quick_start::{my_surface::MySurface, sync_surface::create_new_surface};
use winit::{event::WindowEvent, window::WindowAttributes};
use winit_app::{app_listener::AppWindowEvent, application::Application};

fn launch() -> Result<(), Box<dyn std::error::Error>> {
   let winit_app = Application::new();
   let mut opt_surface: Option<Box<MySurface<'_>>> = None;
   winit_app.run(
       WindowAttributes::default().with_title("wgpu starter app"),
       move |app_window_event| match app_window_event {
           AppWindowEvent::NewWindow(window) => match create_new_surface(window) {
               Ok(value) => {
                   let boxed = Box::new(value);
                   opt_surface = Some(boxed);
               }
               Err(err) => {
                   // warning - Error creating new surface from the window
               }
           },
           AppWindowEvent::OnWindowEvent(event, event_loop) => {
               if let Some(local_surface) = opt_surface.as_mut() {
                   match event {
                       WindowEvent::CloseRequested => {
                           event_loop.exit();
                       }
                       WindowEvent::SurfaceResized(size) => {
                           // Resized
                           local_surface.resize((size.width, size.height));
                       }
                       WindowEvent::RedrawRequested => {
                           match local_surface.get_texture() {
                               Ok((output, view)) => {
                                    let mut encoder = local_surface.on_device(|device| {
                                       device.create_command_encoder(
                                           &wgpu::CommandEncoderDescriptor {
                                               label: Some("Render Encoder"),
                                           },
                                       )
                                   });
                                   {
                                       let _render_pass = encoder.begin_render_pass(
                                           &wgpu::RenderPassDescriptor {
                                               label: Some("render-pass"),
                                               color_attachments: &[Some(
                                                   wgpu::RenderPassColorAttachment {
                                                       view: &view,
                                                       resolve_target: None,
                                                       ops: wgpu::Operations {
                                                           load: wgpu::LoadOp::Clear(
                                                               wgpu::Color {
                                                                   r: 0.2,
                                                                   g: 0.2,
                                                                   b: 0.2,
                                                                   a: 1.0,
                                                               },
                                                           ),
                                                           store: wgpu::StoreOp::Store,
                                                       },
                                                       depth_slice: None,
                                                   },
                                               )],
                                               depth_stencil_attachment: None,
                                               occlusion_query_set: None,
                                               timestamp_writes: None,
                                               multiview_mask: None,
                                           },
                                       );
                                   }
                                   local_surface.submit_to_queue(std::iter::once(encoder.finish()));
                                   output.present();
                               }
                               Err(err) => {
                                   // warning - error creating the texture
                               }
                           }
                       }
                       _ => {}
                   }
               }
           }
       },
   )?;
   Ok(())
}

Implementations

impl<'a> MySurface<'a>

pub async fn new( window: impl Into<SurfaceTarget<'a>>, dimensions: (u32, u32), ) -> GPUStarterResult<Self>

Create a new surface from the given window for the given dimensions (width, height) tuple

pub fn resize(&mut self, (width, height): (u32, u32))

Resizes the given surface with the given dimensions

pub fn get_size(&self) -> (u32, u32)

Retrieve the current size of the underlying configuration

pub fn reconfigure(&self, new_config: &SurfaceConfiguration)

Reconfigure the surface based on updated configuration

pub fn get_texture(&self) -> GPUStarterResult<(SurfaceTexture, TextureView)>

Retrieve a new texture for the current surface

pub fn on_device_mut<R>(&self, fn_on_device: impl FnMut(&Device) -> R) -> R

Execute a function on the device reference stored internally

pub fn on_device<R>(&self, fn_on_device: impl Fn(&Device) -> R) -> R

Execute a function on the device reference stored internally

pub fn on_device_and_texture_mut<R>( &self, fn_device_and_texture: impl Fn(&Device, TextureFormat) -> R, ) -> R

Execute a function on ‘device’ and texture stored internally

pub fn on_device_and_texture<R>( &self, fn_device_and_texture: impl FnMut(&Device, TextureFormat) -> R, ) -> R

Execute a function on ‘device’ and texture stored internally

pub fn on_queue_mut<R>(&self, fn_on_queue: impl FnMut(&Queue) -> R) -> R

Execute a function on the queue reference stored internally

pub fn on_queue<R>(&self, fn_on_queue: impl Fn(&Queue) -> R) -> R

Execute a function on the queue reference stored internally

pub fn on_adapter_mut<R>(&self, fn_on_adapter: impl FnMut(&Adapter) -> R) -> R

Execute a function on the adapter reference stored internally

pub fn on_adapter<R>(&self, fn_on_adapter: impl Fn(&Adapter) -> R) -> R

Execute a function on the adapter reference stored internally

pub fn submit_to_queue<I>(&self, command_buffer: I) -> SubmissionIndex
where I: IntoIterator<Item = CommandBuffer>,

submit the operations on the encoder to the queue created internally.

Queue submission is quite expensive. Should not be used frequently. But, rather batched at the end !