wgpu_quick_start::sync_surface

Function create_new_surface

pub fn create_new_surface<'a>(
    window: Box<dyn Window>,
) -> GPUStarterResult<MySurface<'a>>

This synchronous function helps create a new surface from the given window.

Available with sync feature. (Hence, not available by default).

Can be enabled as below in the crate when being used.

wgpu_quick_start = { version="0.28.3", features = ["sync"] }